Company Features Implemented
This commit is about the implemented company features

Company class and service class has been created to help in the process of logging in and registering.

This commit provide the software the ability to login as a provider and register as a company.

features code has been implemented across all the company section and database connection has been implemented as well